[refactor] : 조회 api 구조 변경
