+++ src/main/java/shop/itbook/itbookshop/pointgroup/pointhistory/service/find/serviceapi/impl/PointHistoryMemberServiceImpl.java
@@ -25,4 +25,6 @@ public class PointHistoryMemberServiceImpl implements PointHistoryMemberService 
         return pointHistoryRepository.findMyPointHistoryListDto(memberNo, pageable,
             pointIncreaseDecreaseContentEnum);
     }
+
+
 }
